[BTAT-10468] Add link to view VAT return on breakdown pages
